backend/requirements.txt
python-dotenv==0.20.0
ldap3==2.9.1
python-dotenv==0.20.0
flask==2.3.3
flask-ldap3-login==0.9.18
Flask-RESTful==0.3.10
orjson==3.6.1
Flask-Login==0.5.0
Flask-HTTPAuth==4.8.0
PyJWT==2.8.0
marshmallow==3.20.2
Flask-Cors==4.0.0